Rename TodoList example file
