+++ examples/bridge/index.ts
@@ -1,3 +1,3 @@
 import "./src/001-HelloWorld";
 import "./src/002-HoverMe";
-import "./src/003-TodoListApp-Client";
+import "./src/003-TodoList";
